cypress: PR feedback
